+++ libs/backend/modules/app/src/use-cases/import-app/test/import-app.api.graphql
@@ -1,0 +1,8 @@
+mutation TestImportApp($input: ImportAppInput!) {
+  importApp(input: $input) {
+    ...TestApp
+    pages {
+      ...PageFull
+    }
+  }
+}
