+++ libs/backend/modules/element/src/test/graphql/TestElement.fragment.graphql.gen.ts
@@ -2,6 +2,7 @@ import * as Types from '@codelab/frontend/abstract/codegen'
 
 import {
   AtomBaseFragment,
+  Export__AtomsFragment,
   AtomFragment,
 } from '../../../../../../frontend/modules/atom/src/graphql/Atom.fragment.graphql.gen'
 import {
@@ -11,6 +12,7 @@ import {
 import { gql } from '@apollo/client'
 import {
   AtomBaseFragmentDoc,
+  Export__AtomsFragmentDoc,
   AtomFragmentDoc,
 } from '../../../../../../frontend/modules/atom/src/graphql/Atom.fragment.graphql.gen'
 import {
