+++ libs/backend/modules/element/src/use-cases/component/create-component/create-component.input.ts
@@ -1,5 +1,8 @@
 import { Field, InputType } from '@nestjs/graphql'
 
+// If you're looking for CreateComponentService - use CreateElementService instead and provide isComponent: true
+// This is just as a facade input for creating an element with component tag
+
 @InputType()
 export class CreateComponentInput {
   @Field()
