+++ libs/backend/modules/element/src/use-cases/component/create-component/index.ts
@@ -1,1 +1,2 @@
 export * from './create-component.input'
+export * from './tests/create-component.api.graphql.gen'
