+++ libs/backend/modules/page/src/use-cases/create-page/create-page.input.ts
@@ -1,3 +1,4 @@
+import { CreateElementChildInput } from '@codelab/backend/modules/element'
 import { Field, InputType } from '@nestjs/graphql'
 
 @InputType()
@@ -7,4 +8,7 @@ export class CreatePageInput {
 
   @Field()
   declare appId: string
+
+  @Field(() => CreateElementChildInput, { nullable: true })
+  declare rootElement?: CreateElementChildInput
 }
