+++ libs/backend/modules/page/src/use-cases/create-page/create-page.service.ts
@@ -5,6 +5,7 @@ import {
   DgraphRepository,
   DgraphUpdateMutationJson,
 } from '@codelab/backend/infra'
+import { CreateElementService } from '@codelab/backend/modules/element'
 import { Injectable } from '@nestjs/common'
 import { Mutation, Txn } from 'dgraph-js-http'
 import { PageValidator } from '../../domain/page.validator'
@@ -12,7 +13,11 @@ import { CreatePageRequest } from './create-page.request'
 
 @Injectable()
 export class CreatePageService extends DgraphCreateUseCase<CreatePageRequest> {
-  constructor(dgraph: DgraphRepository, private pageValidator: PageValidator) {
+  constructor(
+    dgraph: DgraphRepository,
+    private pageValidator: PageValidator,
+    private createElementService: CreateElementService,
+  ) {
     super(dgraph)
   }
 
@@ -27,20 +32,31 @@ export class CreatePageService extends DgraphCreateUseCase<CreatePageRequest> {
     )
   }
 
-  protected createMutation(
-    { input: { appId, name } }: CreatePageRequest,
+  protected async createMutation(
+    { input: { appId, name, rootElement }, currentUser }: CreatePageRequest,
     blankNodeUid: string,
-  ): Mutation {
+  ): Promise<Mutation> {
+    let root: Record<string, any> = {
+      'dgraph.type': [DgraphEntityType.Element],
+      name: 'Root element',
+      children: [],
+      props: '{}',
+    }
+
+    if (rootElement) {
+      const created = await this.createElementService.execute({
+        input: rootElement,
+        currentUser,
+      })
+
+      root = { uid: created.id }
+    }
+
     const createPageJson: DgraphCreateMutationJson<any> = {
       uid: blankNodeUid,
       'dgraph.type': [DgraphEntityType.Page],
       name,
-      root: {
-        'dgraph.type': [DgraphEntityType.Element],
-        name: 'Root element',
-        children: [],
-        props: '{}',
-      },
+      root,
     }
 
     const updateAppJson: DgraphUpdateMutationJson<any> = {
