+++ libs/frontend/modules/page/src/use-cases/create-page/CreatePageForm.tsx
@@ -1,13 +1,12 @@
-import { CreatePageInput } from '@codelab/frontend/abstract/codegen'
 import { FormUniforms } from '@codelab/frontend/view/components'
 import React from 'react'
 import { AutoFields } from 'uniforms-antd'
 import { createPageSchema } from './createPageSchema'
-import { CreatePageFormProps } from './types'
+import { CreatePageFormProps, CreatePageMutationInput } from './types'
 
 export const CreatePageForm = (props: CreatePageFormProps) => {
   return (
-    <FormUniforms<CreatePageInput> schema={createPageSchema} {...props}>
+    <FormUniforms<CreatePageMutationInput> schema={createPageSchema} {...props}>
       <AutoFields omitFields={['appId']} />
     </FormUniforms>
   )
