+++ libs/frontend/modules/page/src/use-cases/create-page/types.ts
@@ -1,7 +1,10 @@
 import { FormUniformsProps } from '@codelab/frontend/view/components'
 import { CreatePageMutationVariables } from '../../graphql/Page.endpoints.graphql.gen'
 
-export type CreatePageMutationInput = CreatePageMutationVariables['input']
+export type CreatePageMutationInput = Omit<
+  CreatePageMutationVariables['input'],
+  'rootElement'
+>
 
 export type CreatePageFormProps = Omit<
   FormUniformsProps<CreatePageMutationInput>,
